+++ app/src/main/java/com/atef/clubhouse/data/remote/feature/channels/mapper/ChannelMapper.kt
@@ -1,11 +1,9 @@
-package com.atef.clubhouse.data.remote.feature.home.mapper
+package com.atef.clubhouse.data.remote.feature.channels.mapper
 
 import com.atef.clubhouse.data.base.mapper.RemoteModelMapper
 import com.atef.clubhouse.data.remote.feature.auth.mapper.ProfileMapper
-import com.atef.clubhouse.data.remote.feature.auth.model.ProfileResponse
-import com.atef.clubhouse.data.remote.feature.home.model.ChannelResponse
-import com.atef.clubhouse.domain.entity.auth.User
-import com.atef.clubhouse.domain.entity.home.Channel
+import com.atef.clubhouse.data.remote.feature.channels.model.ChannelResponse
+import com.atef.clubhouse.domain.entity.channels.Channel
 import javax.inject.Inject
 
 class ChannelMapper @Inject constructor(private val listMapper: ProfileMapper) : RemoteModelMapper<ChannelResponse, Channel> {
