+++ app/src/main/java/com/atef/clubhouse/data/remote/feature/channels/model/ChannelsResponse.kt
@@ -1,6 +1,4 @@
-package com.atef.clubhouse.data.remote.feature.home.model
-
-import com.google.gson.annotations.SerializedName
+package com.atef.clubhouse.data.remote.feature.channels.model
 
 open class ChannelsResponse {
     val channels: List<ChannelResponse>? = null
