+++ app/src/main/java/com/atef/clubhouse/data/remote/feature/channels/model/RuleResponse.kt
@@ -1,7 +1,6 @@
-package com.atef.clubhouse.data.remote.feature.home.model
+package com.atef.clubhouse.data.remote.feature.channels.model
 
 import androidx.annotation.Keep
-import com.google.gson.annotations.SerializedName
 
 @Keep
 data class RuleResponse(
