+++ app/src/main/java/com/atef/clubhouse/data/remote/feature/channels/service/ChannelsApiHandler.kt
@@ -1,7 +1,7 @@
-package com.atef.clubhouse.data.remote.feature.home.service
+package com.atef.clubhouse.data.remote.feature.channels.service
 
 import com.atef.clubhouse.data.remote.base.model.Result
-import com.atef.clubhouse.data.remote.feature.home.model.ChannelsResponse
+import com.atef.clubhouse.data.remote.feature.channels.model.ChannelsResponse
 import retrofit2.http.GET
 import retrofit2.http.Headers
