+++ app/src/main/java/com/atef/clubhouse/data/repository/ChannelsRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.atef.clubhouse.data.repository
 
 import com.atef.clubhouse.data.feature.home.ChannelsRemoteDataSource
 import com.atef.clubhouse.data.remote.base.model.Result
-import com.atef.clubhouse.data.remote.feature.home.mapper.ChannelMapper
-import com.atef.clubhouse.domain.entity.home.Channel
+import com.atef.clubhouse.data.remote.feature.channels.mapper.ChannelMapper
+import com.atef.clubhouse.domain.entity.channels.Channel
 import com.atef.clubhouse.domain.repository.ChannelsRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
