+++ app/src/main/java/com/atef/clubhouse/di/RemoteModule.kt
@@ -8,11 +8,11 @@ import com.atef.clubhouse.data.remote.feature.auth.mapper.ProfileMapper
 import com.atef.clubhouse.data.remote.feature.auth.mapper.UserMapper
 import com.atef.clubhouse.data.remote.feature.auth.model.CompletePhoneNumberAuthResponse
 import com.atef.clubhouse.data.remote.feature.auth.model.ProfileResponse
-import com.atef.clubhouse.data.remote.feature.home.ChannelsRemoteDataSourceImpl
-import com.atef.clubhouse.data.remote.feature.home.mapper.ChannelMapper
-import com.atef.clubhouse.data.remote.feature.home.model.ChannelResponse
+import com.atef.clubhouse.data.remote.feature.channels.ChannelsRemoteDataSourceImpl
+import com.atef.clubhouse.data.remote.feature.channels.mapper.ChannelMapper
+import com.atef.clubhouse.data.remote.feature.channels.model.ChannelResponse
 import com.atef.clubhouse.domain.entity.auth.User
-import com.atef.clubhouse.domain.entity.home.Channel
+import com.atef.clubhouse.domain.entity.channels.Channel
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
