+++ app/src/main/java/com/atef/clubhouse/domain/repository/ChannelsRepository.kt
@@ -1,7 +1,7 @@
 package com.atef.clubhouse.domain.repository
 
 import com.atef.clubhouse.data.remote.base.model.Result
-import com.atef.clubhouse.domain.entity.home.Channel
+import com.atef.clubhouse.domain.entity.channels.Channel
 
 interface ChannelsRepository {
