+++ app/src/main/java/com/atef/clubhouse/ui/home/HomeViewModel.kt
@@ -6,7 +6,7 @@ import com.atef.clubhouse.base.BaseViewModel
 import com.atef.clubhouse.base.Resource
 import com.atef.clubhouse.base.extension.mutable
 import com.atef.clubhouse.domain.entity.auth.User
-import com.atef.clubhouse.domain.entity.home.Channel
+import com.atef.clubhouse.domain.entity.channels.Channel
 import com.atef.clubhouse.domain.feature.auth.LogoutUserCase
 import com.atef.clubhouse.domain.feature.auth.UserUseCase
 import com.atef.clubhouse.domain.feature.home.GetChannelsUseCase
